chore: Little tidy up
